Adjust for better safey with the environment variables.

+++ rootfs/opt/entry-point.sh
@@ -1,13 +1,15 @@
 #!/bin/sh
 set -eu
 
-if [ ! -f $USERLIST_PATH ] ; then
+EMPTY=""
+
+if [ ! -f $USERLIST_PATH -a -n ${DRUPAL_DB_USER:-EMPTY} -a -n ${DRUPAL_DB_PASSWORD:-EMPTY} ] ; then
   cat > $USERLIST_PATH <<EOF
 "$DRUPAL_DB_USER" "md5$(echo -n $DRUPAL_DB_PASSWORD$DRUPAL_DB_USER | md5sum - | cut -d" " -f1)"
 EOF
 fi
 
-if [ -n "$PGHOST" -a -n "$PGDATABASE" -a -n "$PGUSER" -a -n "$PGPASSWORD" ] ; then
+if [ -n ${PGHOST:-EMPTY} -a -n ${PGDATABASE:-EMPTY} -a -n ${PGUSER:-EMPTY} -a -n ${PGPASSWORD:-EMPTY} ] ; then
   ATTEMPTS=0
   while ! pg_isready ; do
     ATTEMPTS=$(($ATTEMPTS + 1))
@@ -21,4 +23,5 @@ if [ -n "$PGHOST" -a -n "$PGDATABASE" -a -n "$PGUSER" -a -n "$PGPASSWORD" ] ; th
 else
   echo "Lacking PG* vars to more generally test the DB connection."
 fi
+
 exec pgbouncer --user=$USER $CONFIG_FILE
